Reintegrate video player into playback history and fix history recording not fired issue

+++ app/src/main/java/com/example/mediaplayerapp/data/playback_history/PlaybackHistoryRepository.java
@@ -68,7 +68,7 @@ public class PlaybackHistoryRepository {
         return recordHistory(mediaUri, false, lastPlaybackPosition);
     }
 
-    public Completable recordVideoHistory(Uri mediaUri, int lastPlaybackPosition) {
+    public Completable recordVideoHistory(Uri mediaUri, long lastPlaybackPosition) {
         return recordHistory(mediaUri, true, lastPlaybackPosition);
     }
